Backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pytz==2021.3
motor==2.5.0
bcrypt==3.2.0
python-jose==3.3.0
email-validator==1.1.3
python-decouple==3.4
pyjwt==2.3.0
python-dateutil==2.8.2
pdfkit==0.6.1
